+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add `Token` presentation
 - Fix `filter_for_quote_tokens` to be more strict (was checking either side of the pair, not quote token)
+- API change `create_single_pair_universe` now to accept `pick_by_highest_vol`
+- API change: Rename `Universe.time_frame` -> `Universe.time_bucket` to be consistent 
 
 # 0.6.9